Restructured game components
