+++ client/src/components/Game/Game.jsx
@@ -1,0 +1,13 @@
+import UserInput from './UserInput/UserInput';
+import Display from './Display/Display'
+
+const Game = () => {
+    return (
+        <>
+            <Display />
+            <UserInput />
+        </>
+    )
+};
+
+export default Game;
