+++ client/src/components/Game/index.js
@@ -1,0 +1,5 @@
+// import Display from './Display/Display';
+// import UserInput from './UserInput/UserInput';
+import Game from  './Game'
+
+export default Game;
